Add files via upload
In this problem, you are given four inputs, the first two numbers are the real and imaginary values ​​of the first complex number, and the next two numbers are the real and imaginary values ​​of the second complex number, respectively. In the resulting output you should:

First line: sum of two given complex numbers

Second line: subtract the value of the second complex number from the value of the first complex number (Z1-Z2)

The third line: the simplified value of the product of two complex numbers

Fourth line: the simplified value of dividing the first complex number by the second complex number

considerations:

Note that a complex number can be negative, positive, decimal, etc.

Note that the number 0 is not shown in the output, that means 0+0.243i is wrong and 0.243i should be printed in the output.

And if both the imaginary part and the real part are equal to zero, display the number 0 in the output.

You must use Structure to store complex numbers.